# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Raptix API
  version: v1
  description: |
    Phase 7A server API. Business endpoints require a verified Bearer OIDC JWT.
    The authenticated principal is derived from the token; actor fields and actor
    headers are not accepted. SSE is advisory progress transport: clients must
    re-read HTTP resources after a reconnect or a cursor gap.
servers:
  - url: /
security:
  - bearerOidc: []
...
paths:
...
  /api/v1/projects:
    get:
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      x-authorization: project membership
      responses:
        '200':
          description: Projects visible to the authenticated principal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectID'
    get:
      operationId: getProject
      x-authorization: project membership
      responses:
        '200':
          description: Project visible to the authenticated principal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/projects/{projectId}/scopes:
    parameters:
      - $ref: '#/components/parameters/ProjectID'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
    get:
      operationId: listProjectScopes
      x-authorization: project membership
      responses:
        '200':
          description: Scopes for a visible project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/projects/{projectId}/runs:
    parameters:
      - $ref: '#/components/parameters/ProjectID'
    get:
      operationId: listProjectRuns
      x-required-capability: run.read
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Runs for a visible project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
    post:
      operationId: createRun
      x-required-capability: run.start
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '201':
          description: Run created, or the original result of an idempotent retry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/runs/{runId}:
    parameters:
      - $ref: '#/components/parameters/RunID'
    get:
      operationId: getRun
      x-required-capability: run.read
      responses:
        '200':
          description: Run view.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/runs/{runId}/tasks:
    parameters:
      - $ref: '#/components/parameters/RunID'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
    get:
      operationId: listRunTasks
      x-required-capability: run.read
      responses:
        '200':
          description: Tasks for a visible run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/runs/{runId}/agents:
    parameters:
      - $ref: '#/components/parameters/RunID'
    get:
      operationId: listRunAgents
      x-required-capability: run.read
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Agents for a visible run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
    post:
      operationId: createRunAgent
      x-required-capability: run.control
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created, or the original result of an idempotent retry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/agents/{agentId}/attempts:
    parameters:
      - $ref: '#/components/parameters/AgentID'
    post:
      operationId: runAgentAttempt
      x-required-capability: run.control
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunAgentAttemptRequest'
      responses:
        '200':
          description: Completed attempt result, or the original result of an idempotent retry. A request timeout does not establish execution success; use the run view and event stream to reconcile state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttemptResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/runs/{runId}/cancel:
    parameters:
      - $ref: '#/components/parameters/RunID'
    post:
      operationId: cancelRun
      x-required-capability: run.control
      responses:
        '200':
          description: Current run state after an idempotent cancellation request. Cancellation never retries execution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/runs/{runId}/evidence:
    parameters:
      - $ref: '#/components/parameters/RunID'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
    get:
      operationId: listRunEvidence
      x-required-capability: evidence.read
      responses:
        '200':
          description: Evidence metadata for a visible run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/evidence/{evidenceId}:
    parameters:
      - $ref: '#/components/parameters/EvidenceID'
    get:
      operationId: getEvidence
      x-required-capability: evidence.read
      responses:
        '200':
          description: Evidence metadata. Storage keys and bytes are never included.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evidence'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/evidence/{evidenceId}/content:
    parameters:
      - $ref: '#/components/parameters/EvidenceID'
    get:
      operationId: downloadEvidenceContent
      x-required-capability: evidence.read
      responses:
        '200':
          description: Existing non-secret artifact content. The response uses the artifact MIME type, Content-Disposition attachment, and Cache-Control no-store; it is never publicly cacheable.
          headers:
            Content-Disposition:
              required: true
              schema:
                type: string
                pattern: '^attachment;'
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/runs/{runId}/findings:
    parameters:
      - $ref: '#/components/parameters/RunID'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
    get:
      operationId: listRunFindings
      x-required-capability: finding.read
      responses:
        '200':
          description: Findings for a visible run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/findings/{findingId}:
    parameters:
      - $ref: '#/components/parameters/FindingID'
    get:
      operationId: getFinding
      x-required-capability: finding.read
      responses:
        '200':
          description: Finding for a visible run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Finding'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/findings/{findingId}/revisions:
    parameters:
      - $ref: '#/components/parameters/FindingID'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
    get:
      operationId: listFindingRevisions
      x-required-capability: finding.read
      responses:
        '200':
          description: Immutable finding revisions, newest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingRevisionList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
    post:
      operationId: reviseFinding
      x-required-capability: finding.review
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviseFindingRequest'
      responses:
        '200':
          description: Finding after the new immutable revision was recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Finding'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/findings/{findingId}/reviews:
    parameters:
      - $ref: '#/components/parameters/FindingID'
    post:
      operationId: reviewFinding
      x-required-capability: finding.review
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewFindingRequest'
      responses:
        '200':
          description: Finding after the review decision. Findings remains the only owner of status transitions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Finding'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/runs/{runId}/reports:
    parameters:
      - $ref: '#/components/parameters/RunID'
    post:
      operationId: createReport
      x-required-capability: report.create
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
      responses:
        '201':
          description: Report request created with an immutable snapshot, or the existing single report for the run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/reports/{reportId}:
    parameters:
      - $ref: '#/components/parameters/ReportID'
    get:
      operationId: getReport
      x-required-capability: report.read
      responses:
        '200':
          description: Report request and publication state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
...
components:
  securitySchemes:
    bearerOidc:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer OIDC JWT verified for issuer, audience, signature, and expiry.
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      description: Page size. Defaults to 50 and may not exceed 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by a prior list response.
      schema:
        type: string
        minLength: 1
        maxLength: 512
...
  responses:
...
    Unauthenticated:
      description: Missing or invalid bearer token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    DependencyUnavailable:
      description: A required dependency is unavailable.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    ProjectList:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Project' }
        nextCursor: { type: [string, 'null'], maxLength: 512 }
...